Extract forward search test logic into test utils
